fix: refactor hooks: useForkRef and useFileInput
