+++ CODEOWNERS
@@ -48,6 +48,7 @@
 /src/components/theme @resure
 
 /src/hooks/useActionHandlers @ogonkov
+/src/hooks/useFileInput @korvin89
 /src/hooks/useOutsideClick @NikitaCG
 /src/hooks/useUniqId @ValeraS
