+++ src/components/Tooltip/Tooltip.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
 import {KeyCode} from '../../constants';
+import {useForkRef} from '../../hooks';
 import {Popup} from '../Popup';
 import type {PopupPlacement} from '../Popup';
 import type {DOMProps} from '../types';
 import {block} from '../utils/cn';
 import {useBoolean} from '../utils/useBoolean';
-import {useForkRef} from '../utils/useForkRef';
 
 import './Tooltip.scss';
