+++ src/components/index.ts
@@ -56,10 +56,7 @@ export * from './utils/useLayer';
 export {Lang, configure} from './utils/configure';
 export * from './utils/useSelect';
 export * from './utils/useListNavigation';
-export * from './utils/useForkRef';
-export * from './utils/setRef';
 export {useOnFocusOutside} from './utils/useOnFocusOutside';
 export * from './utils/interactions';
 export * from './utils/xpath';
-export * from './utils/useFileInput/useFileInput';
 export {getLayersCount} from './utils/LayerManager';
