+++ src/components/utils/FocusTrap.tsx
@@ -3,9 +3,7 @@ import React from 'react';
 import {createFocusTrap} from 'focus-trap';
 import type {FocusTrap as FocusTrapInstance} from 'focus-trap';
 
-import {useUniqId} from '../../hooks';
-
-import {useForkRef} from './useForkRef';
+import {useForkRef, useUniqId} from '../../hooks';
 
 interface FocusTrapContext {
     addNode: (id: string, node: HTMLElement) => void;
