+++ src/hooks/index.ts
@@ -1,5 +1,7 @@
 export * from './useActionHandlers';
 export * from './useBodyScrollLock';
+export * from './useFileInput';
+export * from './useForkRef';
 export * from './useOutsideClick';
 export * from './usePortalContainer';
 export * from './useVirtualElementRef';
