+++ src/hooks/useFileInput/__stories__/UseFileInput.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import type {Meta, StoryFn} from '@storybook/react';
 
-import {Button} from '../../../Button';
+import {Button} from '../../../components/Button';
 import {useFileInput} from '../useFileInput';
 
 export default {title: 'Hooks/useFileInput'} as Meta;
