+++ src/hooks/useFileInput/index.ts
@@ -1,0 +1,2 @@
+export {useFileInput} from './useFileInput';
+export type {UseFileInputProps, UseFileInputOutput, UseFileInputResult} from './useFileInput';
