+++ src/hooks/useForkRef/README.md
@@ -1,0 +1,21 @@
+<!--GITHUB_BLOCK-->
+
+# useForkRef
+
+<!--/GITHUB_BLOCK-->
+
+```tsx
+import {useForkRef} from '@gravity-ui/uikit';
+```
+
+The `useForkRef` hook that can combine refs into a single Ref
+
+## Properties
+
+| Name | Description     |        Type         | Default |
+| :--- | :-------------- | :-----------------: | :-----: |
+| refs | ref-links array | `React.RefObject[]` |         |
+
+## Result
+
+Combined ref. `React.RefCallback`
